Build fixes for onboarding samples in thin base services

+++ build_scripts/build_linux_package.sh
@@ -1,4 +1,4 @@
-#!/bin/bash -ex
+#!/bin/bash -eux
 # Copyright AllSeen Alliance. All rights reserved.
 #
 #    Permission to use, copy, modify, and/or distribute this software for any
@@ -14,11 +14,16 @@
 #    OR IN CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
 #
 # This script collects all the pieces to make the SDK for the cpp builds
-
 env | sort
 
-zip_base=alljoyn-tcl-service-framework-${BUILD_VERSION}-linux-x86_64-sdk-$variant
-sdk_dir=$WORKSPACE/$zip_base
+if [[ -z "${AJ_ROOT+notempty}" ]]; then
+   AJ_ROOT="$(pwd)/../../.."
+fi
+
+[ -d "${WORKSPACE}" ] || { echo "WORKSPACE does not exist"; exit 1; }
+
+zip_base="alljoyn-tcl-service-framework-${BUILD_VERSION}-linux-x86_64-sdk-${variant}"
+sdk_dir="${WORKSPACE}/${zip_base}"
 
 rm -rf $sdk_dir
 mkdir -p $sdk_dir
@@ -28,7 +33,7 @@ mkdir -p $sdk_dir
 ##
 doxyfile=Doxyfile
 
-pushd services/base_tcl
+pushd "${AJ_ROOT}/services/base_tcl"
 #pushd services/base_tcl/$SERVICE # todo
 rm -rf html dist/docs/html
 doxygen $doxyfile ### 2> doxy.out >> /dev/null
@@ -37,6 +42,8 @@ mv -f html dist/docs/html
 cp -rp dist/*    $sdk_dir
 popd
 
+pushd "${WORKSPACE}"
 rm -rf $zip_base.zip
 zip -q -r $zip_base.zip $zip_base
 md5sum $zip_base.zip > md5sum-$zip_base.txt
+popd
